\name{absoluteNormalise.function}
\alias{absoluteNormalise.function}
\title{absoluteNormalise}
\usage{
  absoluteNormalise.function(bead.data, mef.data)
}
\description{
  Absolute normalise to align peaks of bead.data to MEF.
  Returns a list of affine functions from transformed MFI
  relative coordinates to transformed MEF absolute
  coordinates.
}

