\docType{methods}
\name{absoluteNormalise}
\alias{absoluteNormalise}
\title{absoluteNormalise}
\arguments{
  \item{\code{bead.data}}{\link{GatedBeadFlowFrame}}

  \item{mef.data}{\link{data.frame}}
}
\description{
  Absolute normalise to align peaks of bead.data to MEF.
  Returns a list of affine functions from transformed MFI
  relative coordinates to transformed MEF absolute
  coordinates.
}

