\docType{class}
\name{BeadFlowFrame}
\alias{BeadFlowFrame}
\title{BeadFlowFrame}
\usage{
  BeadFlowFrame(fcs.filename, bead.filename)
}
\arguments{
  \item{fcs.filename}{}

  \item{bead.filename}{}
}
\description{
  Extension of \code{\link{flowFrame}} specific for bead
  data.

  constructor
}
\section{Slots}{
  \describe{ \item{\code{beads.mef:} lakdf
  \item{\code{trans}:} aiodnf \item{\code{inv.trans}:}
  adionf }
}
\keyword{beads}
\keyword{flowcytometry}

