\docType{package}
\name{flowBeads-package}
\alias{flowBeads}
\alias{flowBeads-package}
\title{flowBeads}
\description{
  Bioconductor package for working with calibration beads
  in flow cytometry. Based on flowCore package.
}
\author{
  Nikolas Pontikos \email{n.pontikos@gmail.com}
}
\keyword{package}

