\docType{class}
\name{GatedBeadFlowFrame-class}
\alias{GatedBeadFlowFrame}
\alias{GatedBeadFlowFrame-class}
\title{GatedBeadFlowFrame}
\arguments{
  \item{labels}{The resulting labels of the clustering
  assigning each event to a different bead population.}

  \item{clustering.stats}{Three dimensional array
  summarising the stats per channel and population.}

  \item{mef.tranform}{The list of MEF transforms}
}
\description{
  GatedBeadFlowFrame
}

