\docType{methods}
\name{generateReport}
\alias{generateReport}
\alias{generateReport,GatedBeadFlowFrame,character-method}
\title{generateReport}
\arguments{
  \item{bead.data}{\code{\linkS4class{GatedBeadFlowFrame}}}

  \item{output.file}{name of the file to which to output
  the HTML report.}
}
\description{
  Generate an HTML report from a Markdown template using
  \link{knitr}.
}
\seealso{
  knitr
}

