\name{mefTransform}
\alias{mefTransform}
\title{Logicle transformation constructor}
\usage{
  mefTransform(transformationId = "mefTransform", alpha,
    beta)
}
\arguments{
  \item{transformationId}{The name of the transformation.}

  \item{alpha}{The intercept of the MEF transform.}

  \item{beta}{The slope of the MEF transform.}
}
\description{
  Input parameters are to be provided in decades
}

