\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,BeadFlowFrame,character-method}
\alias{plot,BeadFlowFrame,missing-method}
\alias{plot,GatedBeadFlowFrame,character-method}
\title{Plot the results of the clustering.
Plot only the requested channel
which should have a corresponding entry in the MEF files}
\description{
  Plot the results of the clustering. Plot only the
  requested channel which should have a corresponding entry
  in the MEF files

  Ungated bead data, simply draw all channels individually
  (no colours).

  If no argument specified then plot all parameters
}

