\name{relativeNormalise.function}
\alias{relativeNormalise.function}
\title{relativeNormalise}
\usage{
  relativeNormalise.function(bead.data1, bead.data2)
}
\description{
  Relative normalise to align peaks of bead.data1 to those
  of bead.data2 Returns a list of affine functions from
  transformed MFI day 1 coordinates to transformed MFI day
  2 coordinates. This permits comparison of channels across
  two days, provided the detector is stable, even in the
  absence of absolute MEF values.
}

