\docType{methods}
\name{toMEF}
\alias{toMEF}
\alias{toMEF,GatedBeadFlowFrame,flowFrame-method}
\title{toMEF}
\arguments{
  \item{bead.data}{The GatedBeadFlowFrame object containing
  the MEF transform.}

  \item{flow.data}{The flowFrame object on which to apply
  the transform.}
}
\description{
  Given bead.data and a flow.data apply the MEF transform
  to matching channels in \code{flow.data}.
}

