\docType{methods}
\name{absoluteNormalise}
\alias{absoluteNormalise}
\title{absoluteNormalise}
\arguments{
  \item{\code{bead.data}}{\code{\link{GatedBeadFlowFrame}}}

  \item{mef.data}{\link{data.frame}}
}
\value{
  A list of affine functions from transformed MFI relative
  coordinates to transformed MEF absolute coordinates.
}
\description{
  Absolute normalise to align peaks of bead.data to MEF.
}

