\name{mefTransform}
\alias{mefTransform}
\title{Logicle transformation constructor
Input parameters are to be provided in decades}
\usage{
  mefTransform(transformationId = "mefTransform", alpha,
    beta)
}
\description{
  Logicle transformation constructor Input parameters are
  to be provided in decades
}

